iq is not calculated twice
