+++ Software/Development/UARTplot/speedMeasRef.txt
@@ -1,0 +1,300 @@
+r 9
+r 9
+m 9
+m 9
+r 9
+r 9
+m 10
+m 10
+r 9
+r 9
+m 7
+m 7
+r 9
+r 9
+m 3
+m 3
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 1
+m 1
+r 9
+r 9
+m 2
+m 2
+r 9
+r 9
+m 6
+m 6
+r 9
+r 9
+m 10
+m 10
+r 9
+r 9
+m 12
+m 12
+r 9
+r 9
+m 16
+m 16
+r 9
+r 9
+m 19
+m 19
+r 9
+r 9
+m 21
+m 21
+r 9
+r 9
+m 23
+m 23
+r 9
+r 9
+m 22
+m 22
+r 9
+r 9
+m 20
+m 20
+r 9
+r 9
+m 16
+m 16
+r 9
+r 9
+m 12
+m 12
+r 9
+r 9
+m 8
+m 8
+r 9
+r 9
+m 3
+m 3
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 1
+m 1
+r 9
+r 9
+m 3
+m 3
+r 9
+r 9
+m 6
+m 6
+r 9
+r 9
+m 10
+m 10
+r 9
+r 9
+m 14
+m 14
+r 9
+r 9
+m 17
+m 17
+r 9
+r 9
+m 19
+m 19
+r 9
+r 9
+m 20
+m 20
+r 9
+r 9
+m 19
+m 19
+r 9
+r 9
+m 19
+m 19
+r 9
+r 9
+m 18
+m 18
+r 9
+r 9
+m 17
+m 17
+r 9
+r 9
+m 14
+m 14
+r 9
+r 9
+m 10
+m 10
+r 9
+r 9
+m 5
+m 5
+r 9
+r 9
+m 1
+m 1
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 1
+m 1
+r 9
+r 9
+m 2
+m 2
+r 9
+r 9
+m 5
+m 5
+r 9
+r 9
+m 7
+m 7
+r 9
+r 9
+m 11
+m 11
+r 9
+r 9
+m 14
+m 14
+r 9
+r 9
+m 17
+m 17
+r 9
+r 9
+m 21
+m 21
+r 9
+r 9
+m 22
+m 22
+r 9
+r 9
+m 23
+m 23
+r 9
+r 9
+m 21
+m 21
+r 9
+r 9
+m 18
+m 18
+r 9
+r 9
+m 14
+m 14
+r 9
+r 9
+m 9
+m 9
+r 9
+r 9
+m 6
+m 6
+r 9
+r 9
+m 2
+m 2
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
+r 9
+r 9
+m 0
+m 0
