+++ Software/Development/UARTplot/speedMeasRef2.txt
@@ -1,0 +1,664 @@
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 2
+m 2
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 0
+m 0
+r 8
+r 8
+m 1
+m 1
+r 8
+r 8
+m 2
+m 2
+r 8
+r 8
+m 2
+m 2
+r 8
+r 8
+m 3
+m 3
+r 8
+r 8
+m 3
+m 3
+r 8
+r 8
+m 4
+m 4
+r 8
+r 8
+m 5
+m 5
+r 8
+r 8
+m 5
+m 5
+r 8
+r 8
+m 5
+m 5
+r 8
+r 8
+m 5
+m 5
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 10
+m 10
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 9
+m 9
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 6
+m 6
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 7
+m 7
+r 8
+r 8
+m 8
+m 8
